cloud-run-rag-search/1-apps/apps/rag/ingestion/requirements.txt
google-cloud-aiplatform>=1.25.0
google-cloud-bigquery>=3.0.0
google-cloud-storage>=2.16.0
